Refactor: move gpu selection to ui folder
